// apollo-client-for-MyOwn/src/interfaces/interfaces.ts
 
 export default interface ILyndaFriend  {
   allFriends?: any;
   id? :string
   firstName: string
   lastName: string
   email: string
   password?: string
 }
 export interface AddFriendInterface  {
 
  firstName: string
  lastName: string
  email: string
  password:string
}
export interface EditFriendInterface  {
 
  firstName: string
  lastName: string
  email: string
  
}
 
 